# swagger-organizations.yaml
openapi: 3.0.3
info:
  title: Brideside CRM - Organizations API
  version: "1.0.0"
  description: >
    Standalone OpenAPI document covering the Organizations CRUD endpoints.
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:8080
    description: Local development
tags:
  - name: Organizations
    description: Organization management APIs
paths:
  /api/organizations:
    post:
      tags:
        - Organizations
      summary: Create organization
      operationId: createOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationRequest'
            example:
              name: Brideside Chicago
              category: PHOTOGRAPHY
              ownerId: 5
              address: 123 Lake Shore Dr, Chicago, IL
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponseEnvelope'
              example:
                success: true
                message: Organization created
                data:
                  id: 12
                  name: Brideside Chicago
                  category: PHOTOGRAPHY
                  owner:
                    id: 5
                    firstName: Jane
                    lastName: Doe
                    email: jane.doe@brideside.com
                    role: SALES
                    displayName: Jane Doe
                  address: 123 Lake Shore Dr, Chicago, IL
                  createdAt: 2025-11-08T06:30:00Z
                  updatedAt: 2025-11-08T06:30:00Z
    get:
      tags:
        - Organizations
      summary: List organizations
      operationId: listOrganizations
      responses:
        "200":
          description: List of organizations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationListEnvelope'
              example:
                success: true
                message: Organizations fetched
                data:
                  - id: 12
                    name: Brideside Chicago
                    category: PHOTOGRAPHY
                    owner:
                      id: 5
                      firstName: Jane
                      lastName: Doe
                      email: jane.doe@brideside.com
                      role: SALES
                      displayName: Jane Doe
                    address: 123 Lake Shore Dr, Chicago, IL
                    createdAt: 2025-11-08T06:30:00Z
                    updatedAt: 2025-11-08T06:35:00Z
                  - id: 13
                    name: Brideside LA
                    category: MAKEUP
                    owner:
                      id: 8
                      firstName: John
                      lastName: Smith
                      email: john.smith@brideside.com
                      role: CATEGORY_MANAGER
                      displayName: John Smith
                    address: 456 Sunset Blvd, Los Angeles, CA
                    createdAt: 2025-11-08T06:40:00Z
                    updatedAt: 2025-11-08T06:40:00Z
  /api/organizations/categories:
    get:
      tags:
        - Organizations
      summary: List organization categories
      operationId: listOrganizationCategories
      responses:
        "200":
          description: Available categories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryOptionListEnvelope'
              example:
                success: true
                message: Organization categories fetched
                data:
                  - code: PHOTOGRAPHY
                    label: Photography
                  - code: MAKEUP
                    label: Makeup
                  - code: PLANNING_AND_DECOR
                    label: Planning and Decor
  /api/organizations/owners:
    get:
      tags:
        - Organizations
      summary: List eligible owners
      operationId: listOrganizationOwners
      responses:
        "200":
          description: Eligible owner users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerOptionListEnvelope'
              example:
                success: true
                message: Organization owners fetched
                data:
                  - id: 5
                    firstName: Jane
                    lastName: Doe
                    email: jane.doe@brideside.com
                    role: SALES
                    displayName: Jane Doe
                  - id: 8
                    firstName: John
                    lastName: Smith
                    email: john.smith@brideside.com
                    role: CATEGORY_MANAGER
                    displayName: John Smith
  /api/organizations/{id}:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
    get:
      tags:
        - Organizations
      summary: Get organization
      operationId: getOrganization
      responses:
        "200":
          description: Organization fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponseEnvelope'
        "404":
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Organizations
      summary: Update organization
      operationId: updateOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationRequest'
            example:
              name: Brideside Downtown
              category: PLANNING_AND_DECOR
              ownerId: 5
              address: 789 Michigan Ave, Chicago, IL
      responses:
        "200":
          description: Organization updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponseEnvelope'
        "404":
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Organizations
      summary: Delete organization
      operationId: deleteOrganization
      responses:
        "200":
          description: Organization deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyEnvelope'
              example:
                success: true
                message: Organization deleted
                data: null
        "404":
          $ref: '#/components/responses/NotFound'
components:
  parameters:
    OrganizationId:
      name: id
      in: path
      description: Organization identifier
      required: true
      schema:
        type: integer
        format: int64
  responses:
    NotFound:
      description: Organization not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            message: Organization not found with id 999
            data: null
  schemas:
    OrganizationRequest:
      type: object
      required:
        - name
        - category
      properties:
        name:
          type: string
          maxLength: 255
          description: Organization name
        category:
          type: string
          description: Organization category enum value
          enum:
            - PHOTOGRAPHY
            - MAKEUP
            - PLANNING_AND_DECOR
        ownerId:
          type: integer
          format: int64
          description: User id for the organization owner (SALES or CATEGORY_MANAGER)
        address:
          type: string
          maxLength: 500
          description: Organization address
    OrganizationResource:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        category:
          type: string
          enum:
            - PHOTOGRAPHY
            - MAKEUP
            - PLANNING_AND_DECOR
        owner:
          $ref: '#/components/schemas/OwnerSummary'
        address:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OwnerSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum:
            - SALES
            - CATEGORY_MANAGER
        displayName:
          type: string
    OwnerOptionListEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/OwnerSummary'
    CategoryOption:
      type: object
      properties:
        code:
          type: string
        label:
          type: string
    CategoryOptionListEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/CategoryOption'
    ApiEnvelope:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          nullable: true
    OrganizationResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/OrganizationResource'
    OrganizationListEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/OrganizationResource'
    EmptyEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              nullable: true
              default: null
    ErrorEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              nullable: true
              default: null